requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
redis==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiortc==1.6.0
sqlmodel==0.0.8
sqlalchemy==2.0.19
httpx==0.24.1